rename internal/pipelines → internal/pipeline
Follows singular-name convention from CLAUDE.md.
